Refactored OgcService directory, split into single top-level components.
