+++ frontend/src/Components/FeatureTypeTable/FeatureTypeTable.tsx
