+++ frontend/src/Components/LayerTable/LayerTable.tsx
