+++ frontend/src/Components/WfsTable/WfsTable.tsx
